Initial dump of backend functionality (#6)
* feat: Create basic networks for validation in TF and Keras

The networks can be used to validate our networks in the future, 
although they should probably have more layers added to them since they 
are very simple in their current state.

* feat: Begin first pass at backend functionality and classes

This commit includes two things: the first is the beginnings of the 
control script, and the second is some initial designs for Layer 
classes.

The control script may get broken up and put into TKinter functions, but 
having it all in one place right now helps me plan it out better.

The Layer class and its children are currently very basic because I'm 
still trying to figure out how it's going to work, but I don't think 
they will need to be all that complicated in the long run since Keras 
abstracts a lot of it for us. I still need to find a good design for the 
parent class that makes inheritence useful.

* chore: Create tests

Mocking needs to be added to `test_create_network()` so it doesn't 
produce a file every time. The last line of the function can be 
uncommented to see an example of a generated neural network.

* chore: Add `__pycache__` to .gitignore

* feat: Flesh out control script and Layer class

The `Layer` class is now much closer to where it needs to be. It will 
just take a little tweaking to make sure it can be integrated into the 
front end.

The control script can read CSV files and generate the neural network 
file. However, I ran into some problems trying to import that file so 
the network can be trained. I will most likely need some help with that.

* feat: Break out Layer class into child classes

the `Layer` class is now the parent class `GenericLayer` that is 
inherited by other layers. This should make it more clear how to create 
new layers.

* feat: Finish base tests with mocks

Both tests are now running with mocks so no files are actually written.

* fix: Resolve import errors from importing created network

The network functions can now be imported once the file has been written 
by `generate_network()`.