Backend working (#19)
* feat: Begin first pass at backend functionality and classes

This commit includes two things: the first is the beginnings of the 
control script, and the second is some initial designs for Layer 
classes.

The control script may get broken up and put into TKinter functions, but 
having it all in one place right now helps me plan it out better.

The Layer class and its children are currently very basic because I'm 
still trying to figure out how it's going to work, but I don't think 
they will need to be all that complicated in the long run since Keras 
abstracts a lot of it for us. I still need to find a good design for the 
parent class that makes inheritence useful.

* chore: Create tests

Mocking needs to be added to `test_create_network()` so it doesn't 
produce a file every time. The last line of the function can be 
uncommented to see an example of a generated neural network.

* feat: Flesh out control script and Layer class

The `Layer` class is now much closer to where it needs to be. It will 
just take a little tweaking to make sure it can be integrated into the 
front end.

The control script can read CSV files and generate the neural network 
file. However, I ran into some problems trying to import that file so 
the network can be trained. I will most likely need some help with that.

* feat: Break out Layer class into child classes

the `Layer` class is now the parent class `GenericLayer` that is 
inherited by other layers. This should make it more clear how to create 
new layers.

* feat: Finish base tests with mocks

Both tests are now running with mocks so no files are actually written.

* fix: Resolve import errors from importing created network

The network functions can now be imported once the file has been written 
by `generate_network()`.

* feat: Add input layer

Input layers are just Dense layers with an input shape specified. To 
make the division more clear a separate class was created.

* test: Add tests for property setting

* feat: Implement property checks, make functions thread safe

This commit adds checks to the `canvas_properties` values to handle any 
errors and prevent crashes later on in execution.

It also adds a private thread lock to prevent the functions 
`generate_network()` and `train_network()` from being run 
simultaneously.

* feat: Add signal handling for graceful termination

Using a signal handler, a thread that is terminated will now release 
`thread_lock` if it has been acquired. This prevents the lock from 
remaining locked after the thread has been terminated.

* style: Change `properties` to `layer_properties`

* fix: Test end-to-end functionality and organize workspace

This commit is a little messy so sorry for that. As of now, end-to-end control functionality is working as expected. This can be tested by navigating to the project root and running `python -m tests.test_end_to_end`, which will unzip the MNIST data, generate a model, and train it on the data. The model should be about 92% accurate, which isn't great but it's a super simple network so there is a lot of room for improvement.

At this point, the project should bbe ready to integrate the front end and back end. There are most likely still some details that need to be ironed out, but they may not be apparent until the integration is underway.

* chore: Update README

This needs to be updated further, but at least it's not totally useless
now.

* chore: Clean up imports and PEP8 style