Training in separate process (#36)
* fix: Make tweaks to get end-to-end working

Tweaks to dictionary keys and sanity checks were made to so the program does not error out. Some of these will need to be reversed or fixed again once we fully integrate everything but they should be marked.

As of now, the system works - a script can be generated and a model can be trained - however, threading is not working as expected and will need some work done on it. This mostly affects the ability to use the GUI while training is happening, and the user cannot cancel training using the button because of this.

* fix: Add Input layer correctly

Input layers cannot be added using the traditional `model.add()` and must be passed directly to the model using the `__call__()` method. The Keras baseline network has been updated to reflect this so it is clear and can be verified.

* chore: Add `to_categorical()` and update logging

To use some loss functions, the labels need to be converted to categorical labels using the `to_categorical()` function. This will be added to the generated script if a loss function other than `sparse_categorical_crossentropy` is chosen.

This commit also updates logging to log to a file instead of the console.

* feat: Perform training in new process

This commit adds a new class and three new functions:
- `ControlProcess`
- `train_in_new_thread()`
- `terminate_training()`
- `__create_new_process()`
which allow training to be done in a new process leaving the main process (GUI) free.

`train_in_new_thread()` is what actually starts the process; it first checks if the process is already running, and attempts to join it if not. This is done to see if a process has already run successfully, and if so, a new one will be created. This prevents a new process from being created the first time or if the previous process was terminated, since `terminate_training()` creates a new process automatically.

The `ControlProcess` class is a wrapper around `multiprocessing.Process` that performs some exception handling. This is mostly redundant since the functions also perform similar checks, but it makes sure that exceptions are not thrown if a process is started more than once or terminated without running.

* chore: Update checks and comments

Some of the checks that were being handled by the front end are no longer being handled or are going to be phase out soon, so they were re-added to `set_properties()`

* style: Make small tweaks and add comments

Tests were also updated, although they do not currently work because they are dependent on the `properties` dictionary structure, which is still in flux.

* tests: Update `canvas_properties` keys and fix tests

* feat: Establish connection between processes for sending status information

Problem:
The original idea for updating the status box from a child process was to simply have the `add_text()` method be called from the child process. Unfortunately, this does not call the method in the parent process.

Solution:
`multiprocessing.Pipe` can be used to send information between parent and child processes. Using this, the child sends status messages to the parent process and the TKinter window checks for messages on an interval using the `after()` method.

Notes:
If a backlog of messages were to build up, theoretically one would only be shown every 500ms, which is what the interval is set to. Because this is only used for the training method, it's unlikely that this would ever happen, or at most 1-2 messages would queue up and be delivered at a delay. Fortunately, the training should be expected to take long enough that this would be relatively insignificant.

* fix: Update property limits

The minimum of slots was set in the SDS to be 3, so this could probably be changed back, however a network with just an input and an output is a valid network and it was bugging me so I changed it for the time being.

The maximum layer dimensions was set at 100 but the MNIST dataset dimensions are 784, so we should discuss what the new max should be. I set it to 1000 but I don't think it's unheard of to have several thousand features.